Adding support for the git-review command line tool
The OpenStack doc workflow (and entire gerrit workflow) has the
option of using a tool called git-review that automates ChangeID:
creation and management,automates rebasing of local commits
as well as simplifies submiting to gerrit and pulling and
pushing reviews.

Usage of the tool is entirely optional, however it does require
a pointer file named .gitreview in the root of the repository
that is referenced.

I am totally new to this project, if you guys don't want to merge
this then no big deal. However I think this will help with dev's
coming over from OpenStack as it will provide a optional workflow
that is closer to the OpenStack workflow.

It is necessary to run git review -s to append the change-id into
git-hooks the first time you use the tool

Change-Id: I9cec21ae32f30b1a54bb430efc5be9d929cfd70d
Signed-off-by: Colin McNamara <colin@2cups.com>