Update patch set 1
Patch Set 1:

> To Thanh,
 > I see. (I removed all code because code huge revision, almost
 > entirely new codes. The removal represents a reset. Now I revert
 > this commit)

The purpose of version control is to track changes in the code. Even if you rewrote much of the code you should be updating the existing code base so that your Version Control system can properly track the history of changes to the files. Wiping out your code every milestone defeats the purpose of maintaining code control in Git.

Ideally though you should be breaking your changes into features / smaller components so that they make it into Gerrit in sizeable and meaningful contributions that can be reviewed and history captured of the code changes. Providing a code dump each release is not the correct way to develop in open source.

Patch-set: 1