Update patch set 1
Patch Set 1:

Thanks, Thanh! We'll do our best to follow the code version control principle. snmp4sdn code dump is due to internal development reason that we needed custom API in ad-sal, but the modified ad-sal could not be in opendaylight git repo, so for a long time we did not update code in snmp4sdn's git repo. :( Now this situation is freed. So snmp4sdn will be updated more frequently.

Patch-set: 1