Update patch set 1
Patch Set 1:

There will be some problems if we use the oid which consists of arpTableEntryPhyAddrOID, midStuffForArpTableEntryOID and ipAddress to get the mac address corresponding to the ipAddress in arp table.


Because the midStuffForArpTableEntryOID is not always 5121 for all devices which come from different vendors.


Furthermore, the midStuffForArpTableEntryOID is also not a constant even for a same device. For example, when I use MG-SOFT MIB Browser software to get ipNetToMediaPhysAddress(this oid is arpTableEntryPhyAddrOID, and it is 1.3.6.1.2.1.4.22.1.2) from h3c's device, the result is as follows.


***** SNMP QUERY STARTED *****
1: ipNetToMediaPhysAddress.37.192.168.1.1 (PhysAddress) 70:F9:6D:EB:03:17 [70.F9.6D.EB.03.17 (hex)]
2: ipNetToMediaPhysAddress.39.192.168.5.1 (PhysAddress) 70:F9:6D:EB:03:17 [70.F9.6D.EB.03.17 (hex)]
3: ipNetToMediaPhysAddress.47.172.168.2.1 (PhysAddress) 00:50:56:60:AC:56 [Vmware_60:AC:56]
4: ipNetToMediaPhysAddress.26446.98.0.0.1 (PhysAddress) 00:0F:E1:21:21:15 [Id_21:21:15]
5: ipNetToMediaPhysAddress.26446.98.0.0.5 (PhysAddress) 5C:DD:70:0F:00:D0 [5C.DD.70.0F.00.D0 (hex)]


We can see that the midStuffForArpTableEntryOID is not a constant even for a same device from this example.

Patch-set: 1
Reviewer: Gerrit User 5002 <5002@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0