Move ad-sal to snmp4sdn
In Boron, felix.dependencymanager (felix.dm) of version 4 is used but conflicts with sal version 0.7.0. which uses felix.dm of version 3.
Other modifications to ad-sal also might be required in the future, so we move ad-sal code to snmp4sdn.

In pom files the version felix.dm just follows odlparent.

Change-Id: Ib84051a8c95141cbaf7ff9e86f9d304cd982bbf1
Signed-off-by: ylhsieh <ylhsieh@itri.org.tw>