Fix for snmp4sdn Karaf 4 migration (2nd fix)
This gerrits continues from
https://git.opendaylight.org/gerrit/#/c/55185

and then:

(1)
SingleFeatureTest was skipped and is included now.
For the two feature, feature-snmp4sdn (for karaf 3) and feature4-snmp4sdn (for karaf 4),
in pom.xml we removed the <build> section so that it will
be built according to the parent, ODL suggested odlparent.
That is, SingleFeatureTest would be executed.

(2)
Fix a SingleFeatureTest failure cause, referring to
the wiki Karaf_4_Migration_Guide provided Caveats-Blueprint

(3)
Fix a SingleFeatureTest failure cause, we removed dependency to
org.eclipse.tycho:org.eclipse.osgi .

(4)
Following (3) removal, however, the library in it, such as
org.eclipse.osgi.framework.console.*, is removed,
in which snmp4sdn uses CommandInterpreter and CommandProvider
(though actually we don't need the code now).
So we workaround this by adding our own Classes, CommandInterpreter and CommandProvider,
to replace the original ones. We focus on SingleFeatureTest fixing in this gerrit,
we'll do housekeeping later.

Change-Id: I0b8e673b747972e7486af88a4cc246fe602206aa
Signed-off-by: ylhsieh <ylhsieh@itri.org.tw>