Add Karaf build profile
Project local Karaf distros are handy for devs to test their work
however is unneeded by autorelease builds and should not be released
as part of the Simultanious Release. Add a profile that is active by
default so that default behaviour is unchanged however allows the
autorelease project to disable building this module.

snmp4sdn appears to have already had distribution-karaf disabled so
leaving the default to disabled to keep the same functionality.

Issue: RELENG-86
Change-Id: I2bc277b18768227dcc926ef4c477b27db00556a0
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>